add option -fasta from local file system
