+++ dashsite/pang/graph/PangraphBuilder.py
@@ -1,0 +1,12 @@
+import abc
+
+from metadata.MultialignmentMetadata import MultialignmentMetadata
+
+
+class PangraphBuilder(abc.ABC):
+    def __init__(self, genomes_info: MultialignmentMetadata):
+        self.sequences_names = genomes_info.get_all_mafnames()
+
+    @abc.abstractmethod
+    def build(self, input, pangraph):
+        pass
