+++ tests/Files/seq/fasta/seq1.fasta
@@ -1,0 +1,2 @@
+>seq1
+ACTGGCGTAC
