+++ tests/Files/seq/fasta/seq2.fasta
@@ -1,0 +1,2 @@
+>seq2
+GTACCGTACC
