+++ tests/Files/seq/fasta/seq3.fasta
@@ -1,0 +1,2 @@
+>seq3
+TTAGCGCACG
