+++ tests/Files/seq/seq_metadata.json
@@ -1,0 +1,10 @@
+{"title": "Ebola",
+  "source": "Ebola portal",
+  "data": [
+      {"genbankID": "", "assemblyID": "", "mafname": "seq0", "name": "seq0", "group": "1"},
+      {"genbankID": "", "assemblyID": "", "mafname": "seq1", "name": "seq1", "group": "1"},
+      {"genbankID": "", "assemblyID": "", "mafname": "seq2", "name": "seq2", "group": "2"},
+      {"genbankID": "", "assemblyID": "", "mafname": "seq3", "name": "seq3", "group": "2"},
+      {"genbankID": "", "assemblyID": "", "mafname": "seq4", "name": "seq4", "group": "2"}
+    ]
+}
